Playlist loads are asynchronous
Along with some restructuring of 3p assets, the main thrust of this
change was to make playlist load asynchronously.  This means fetching
the playlist via ajax and introducing some client-side templating.

Normally, for client-side templating, I use template() underscore.js,
but this time, I wrote my own.  It's more verbose than underscore, but
it is entirely declarative.  Templates are simple <TEMPLATE> tags.
There there is no evaluating javascript within the template.  It's
just CSS selectors to replace attributes and text.

I am only moderately happy with the UI.  It is functional, which is my
personal goal for this project.