added "100 Random Tracks" and playlist shuffle
This patch adds a button to the top-level nav called
"100 Random Tracks".  When pressed, 100 randomly selected (but unique)
tracks are added to the current playlist.

On the playlist view, a new button called "shuffle" will shuffle the
order of the current playlist.  Note that this shuffling is not
persisted to the server.